Move GUIMessageQueue to separate file
