+++ src/Engine/Objects/NPC.cpp
@@ -14,6 +14,7 @@
 #include "Engine/Tables/NPCTable.h"
 
 #include "GUI/GUIButton.h"
+#include "GUI/GUIMessageQueue.h"
 #include "GUI/UI/UIHouses.h"
 #include "GUI/UI/UIStatusBar.h"
 #include "GUI/UI/UITransition.h"
