+++ src/GUI/GUIEnums.h
@@ -264,6 +264,44 @@ enum MENU_STATE : int32_t {
     MENU_NONE = -1,
 };
 
+enum class CURRENT_SCREEN {
+    SCREEN_GAME = 0,
+    SCREEN_MENU = 1,
+    SCREEN_OPTIONS = 2,
+    SCREEN_BOOKS = 3,
+    SCREEN_NPC_DIALOGUE = 4,
+    SCREEN_REST = 5,
+    SCREEN_6 = 6,
+    SCREEN_CHARACTERS = 7,
+    SCREEN_SPELL_BOOK = 8,
+    SCREEN_CREATORS = 9,
+    SCREEN_CHEST = 10,
+    SCREEN_SAVEGAME = 11,
+    SCREEN_LOADGAME = 12,
+    SCREEN_HOUSE = 13,
+    SCREEN_SHOP_INVENTORY = 14,
+    SCREEN_CHEST_INVENTORY = 15,
+    SCREEN_VIDEO = 16,
+    SCREEN_CHANGE_LOCATION = 17,
+    SCREEN_INPUT_BLV = 18,
+    SCREEN_BRANCHLESS_NPC_DIALOG = 19,
+    SCREEN_20 = 20,
+    SCREEN_PARTY_CREATION = 21,
+    SCREEN_GAMEOVER_WINDOW = 22,
+    SCREEN_CASTING = 23,
+    SCREEN_24 = 24,
+    SCREEN_19 = 25,
+    SCREEN_KEYBOARD_OPTIONS = 26,
+    SCREEN_1B = 27,
+    SCREEN_VIDEO_OPTIONS = 28,
+
+    SCREEN_63 = 0x63,
+    SCREEN_64 = 0x64,
+    SCREEN_67 = 0x67,
+    SCREEN_QUICK_REFERENCE = 0x68,
+
+    SCREEN_DEBUG = 999,
+};
 
 enum WindowType {
     WINDOW_null = 0,
