+++ src/GUI/UI/Houses/Bank.cpp
@@ -3,6 +3,7 @@
 #include "GUI/UI/Houses/Bank.h"
 
 #include "GUI/UI/UIStatusBar.h"
+#include "GUI/GUIMessageQueue.h"
 #include "GUI/GUIFont.h"
 #include "GUI/GUIButton.h"
