+++ src/GUI/UI/Houses/MercenaryGuild.cpp
@@ -1,6 +1,7 @@
 #include "GUI/UI/Houses/MercenaryGuild.h"
 
 #include "GUI/UI/UIStatusBar.h"
+#include "GUI/GUIMessageQueue.h"
 #include "GUI/GUIFont.h"
 #include "GUI/GUIButton.h"
