+++ src/GUI/UI/Houses/Tavern.cpp
@@ -5,6 +5,7 @@
 #include "GUI/UI/UIStatusBar.h"
 #include "GUI/GUIFont.h"
 #include "GUI/GUIButton.h"
+#include "GUI/GUIMessageQueue.h"
 
 #include "Engine/Localization.h"
 #include "Engine/PriceCalculator.h"
