+++ src/GUI/UI/Houses/Temple.cpp
@@ -6,6 +6,7 @@
 #include "GUI/UI/UIGame.h"
 #include "GUI/GUIFont.h"
 #include "GUI/GUIButton.h"
+#include "GUI/GUIMessageQueue.h"
 
 #include "Engine/Localization.h"
 #include "Engine/PriceCalculator.h"
