+++ src/GUI/UI/UIArena.cpp
@@ -16,6 +16,7 @@
 #include "GUI/GUIFont.h"
 #include "GUI/GUIWindow.h"
 #include "GUI/GUIButton.h"
+#include "GUI/GUIMessageQueue.h"
 #include "GUI/UI/UIHouses.h"
 
 #include "Media/Audio/AudioPlayer.h"
