+++ src/GUI/UI/UICharacter.cpp
@@ -21,6 +21,7 @@
 #include "GUI/GUIWindow.h"
 #include "GUI/GUIButton.h"
 #include "GUI/GUIFont.h"
+#include "GUI/GUIMessageQueue.h"
 #include "GUI/UI/UIInventory.h"
 
 #include "Io/Mouse.h"
