+++ src/GUI/UI/UIGame.cpp
@@ -39,13 +39,13 @@
 #include "GUI/GUIButton.h"
 #include "GUI/GUIFont.h"
 #include "GUI/GUIWindow.h"
+#include "GUI/GUIMessageQueue.h"
 #include "GUI/UI/UICharacter.h"
 #include "GUI/UI/UIHouses.h"
 #include "GUI/UI/UIStatusBar.h"
 
 #include "Io/InputAction.h"
 #include "Io/Mouse.h"
-
 
 #include "Utility/Math/TrigLut.h"
