+++ src/GUI/UI/UIMainMenu.cpp
@@ -7,6 +7,7 @@
 
 #include "GUI/GUIButton.h"
 #include "GUI/GUIFont.h"
+#include "GUI/GUIMessageQueue.h"
 
 #include "Io/Mouse.h"
