+++ src/GUI/UI/UISaveLoad.cpp
@@ -21,6 +21,7 @@
 
 #include "GUI/GUIButton.h"
 #include "GUI/GUIFont.h"
+#include "GUI/GUIMessageQueue.h"
 
 using Io::TextInputType;
